pyproject.toml
[tool.poetry]
name = "starlette_demo"
version = "0.1.0"
description = "A demo reimplementing the snippets tutorial of Django Rest Framework"
authors = ["lewoudar <lewoudar@gmail.com>"]
license = "MIT"

packages = [
    {include = "pastebin"}
]

[tool.poetry.scripts]
"init-db" = 'pastebin.scripts:initialize_database'

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.13.4"
sqlalchemy = "^1.3.17"
sqlalchemy-utils = "^0.36.6"
"zope.sqlalchemy" = "^1.3"
uvicorn = "^0.11.5"
click = "^7.1.2"
alembic = "^1.4.2"
bcrypt = "^3.1.7"
colorama = "^0.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-cov = "^2.9.0"
httpie = "^2.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
